+++ src/pages/CallPage/CallPage.tsx
@@ -17,10 +17,8 @@ import {
 import { useAudioCall } from '../../../src/hooks/useAudioCall';
 import { CallStatus } from '../../../src/hooks/useAudioCall';
 import Header from '../../../src/components/header/header';
-import { ProgressBar } from 'react-native-paper';
 import palette from '../../../src/assets/styles/theme';
 import IconButton from '../../../src/components/icon-button/IconButton';
-import Icon from '../../../src/components/icons/icons';
 import { setAudioReceiveHandler } from './socketManager';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 import { AudioBars } from './MicLevelBar';
@@ -28,307 +26,9 @@ import Purchases from 'react-native-purchases';
 import { getCurrentOffering, updatePurchaseStatus } from '../../services/inappService';
 import { getRemainingTime } from '../../apis/voice';
 import { getUserNickname } from '../../utils/storageUtils';
-// 결제 모달 컴포넌트
-const PaymentModal: React.FC<{
-  visible: boolean;
-  onClose: () => void;
-  onPayment: (amount: number) => void;
-}> = ({ visible, onClose, onPayment }) => {
-  const paymentOptions = [
-    { minutes: 10, price: 2200, label: '10분' },
-    { minutes: 30, price: 5900, label: '30분' },
-    { minutes: 60, price: 10900, label: '60분' },
-    { minutes: 120, price: 19900, label: '120분' },
-  ];
-
-  return (
-    <Modal animationType="slide" transparent={true} visible={visible} onRequestClose={onClose}>
-      <View
-        style={{
-          flex: 1,
-          backgroundColor: 'rgba(0, 0, 0, 0.5)',
-          justifyContent: 'flex-end',
-        }}>
-        <View
-          style={{
-            backgroundColor: 'white',
-            borderTopLeftRadius: 20,
-            borderTopRightRadius: 20,
-            paddingTop: 20,
-            paddingHorizontal: 20,
-            paddingBottom: 40,
-            minHeight: 400,
-          }}>
-          <View
-            style={{
-              flexDirection: 'row',
-              justifyContent: 'space-between',
-              alignItems: 'center',
-              marginBottom: 20,
-            }}>
-            <Text
-              style={{
-                fontSize: 20,
-                fontWeight: 'bold',
-                color: palette.neutral[900],
-              }}>
-              통화 시간 충전
-            </Text>
-            <TouchableOpacity
-              onPress={onClose}
-              style={{
-                padding: 8,
-              }}>
-              <Text
-                style={{
-                  fontSize: 24,
-                  color: palette.neutral[500],
-                }}>
-                ✕
-              </Text>
-            </TouchableOpacity>
-          </View>
-
-          <Text
-            style={{
-              fontSize: 16,
-              color: palette.neutral[500],
-              marginBottom: 30,
-            }}>
-            쿠키에게 전화를 하려면 통화 시간권 구입이 필요해요
-          </Text>
-
-          <View style={{ gap: 12 }}>
-            {paymentOptions.map((option, index) => (
-              <TouchableOpacity
-                key={index}
-                style={{
-                  flexDirection: 'row',
-                  justifyContent: 'space-between',
-                  alignItems: 'center',
-                  padding: 16,
-                  borderRadius: 12,
-                  backgroundColor: palette.neutral[100],
-                  marginBottom: 12,
-                }}
-                onPress={() => {
-                  console.log(`${option.label} 결제 시작 - ${option.price}원`);
-                  onPayment(option.minutes);
-                  onClose();
-                }}>
-                <Text
-                  style={{
-                    fontSize: 18,
-                    fontWeight: '600',
-                    color: palette.neutral[900],
-                  }}>
-                  {option.label}
-                </Text>
-                <Text
-                  style={{
-                    fontSize: 16,
-                    color: palette.primary[500],
-                    fontWeight: '500',
-                  }}>
-                  {option.price.toLocaleString()}원
-                </Text>
-              </TouchableOpacity>
-            ))}
-          </View>
-        </View>
-      </View>
-    </Modal>
-  );
-};
-
-const CallTimer: React.FC<{
-  remainingTime: number;
-  totalTime: number; // 전체 시간 (선택적)
-  onChargePress: () => void;
-  isLoading?: boolean; // 로딩 상태 추가
-}> = ({ totalTime, remainingTime, onChargePress, isLoading = false }) => {
-  //console.log('CallTimer 렌더링', { remainingTime, totalTime });
-
-  // 초를 hh:mm:ss 형식으로 변환하는 함수
-  const formatTime = (seconds: number): string => {
-    const hours = Math.floor(seconds / 3600);
-    const minutes = Math.floor((seconds % 3600) / 60);
-    const secs = seconds % 60;
-
-    return `${hours.toString().padStart(2, '0')}:${minutes.toString().padStart(2, '0')}:${secs.toString().padStart(2, '0')}`;
-  };
-
-  const isCritical = remainingTime <= 180; // 3분 이하일 때 critical로 간주
-  const color = isCritical ? '#DA1E28' : '#8CC1FF';
-  const progress = totalTime > 0 ? remainingTime / totalTime : 0;
-
-  return (
-    <View style={{ borderColor: 'red', flexDirection: 'row', marginTop: 36, gap: 8 }}>
-      <View
-        style={{
-          width: 40,
-          height: 40,
-          borderRadius: 20,
-          backgroundColor: color,
-          justifyContent: 'center',
-          alignItems: 'center',
-        }}>
-        <Icon name="clock" width="24" height="24" color={palette.neutral[50]} />
-      </View>
-      <View
-        style={{
-          flexDirection: 'column',
-          justifyContent: 'space-between',
-          borderColor: 'gray',
-          flex: 1,
-        }}>
-        <View style={{ flexDirection: 'row', justifyContent: 'space-between' }}>
-          <View
-            style={{
-              flexDirection: 'row',
-              gap: 8,
-              justifyContent: 'space-between',
-              alignItems: 'flex-end',
-              flex: 1,
-            }}>
-            <View style={{ flexDirection: 'row', gap: 8, alignItems: 'flex-end' }}>
-              {isLoading ? (
-                <View style={{ flexDirection: 'row', alignItems: 'center', gap: 8 }}>
-                  <ActivityIndicator size="small" color="white" />
-                  <Text
-                    style={{
-                      color: 'white',
-                      fontSize: 14,
-                      fontFamily: 'Pretendard-Medium',
-                    }}>
-                    로딩중...
-                  </Text>
-                </View>
-              ) : (
-                <>
-                  <Text
-                    style={{
-                      color: 'white',
-                      fontSize: 18,
-                      lineHeight: 28,
-                      fontFamily: 'Pretendard-Medium',
-                      width: 85,
-                    }}>
-                    {formatTime(remainingTime)}
-                  </Text>
-                  <Text
-                    style={{
-                      color: 'white',
-                      fontSize: 12,
-                      fontFamily: 'Pretendard-Medium',
-                    }}>
-                    남았습니다
-                  </Text>
-                </>
-              )}
-            </View>
-            <TouchableOpacity
-              onPress={onChargePress}
-              disabled={isLoading}
-              style={{
-                backgroundColor: isLoading ? palette.neutral[400] : palette.primary[500],
-                padding: 5,
-                borderRadius: 5,
-              }}>
-              <Text style={{ color: 'white', fontSize: 12, fontFamily: 'Pretendard-SemiBold' }}>
-                충전하기
-              </Text>
-            </TouchableOpacity>
-          </View>
-        </View>
-        <ProgressBar
-          progress={progress}
-          color={color}
-          style={{ height: 8, backgroundColor: '#E0E0E0', borderRadius: 10 }}
-        />
-      </View>
-    </View>
-  );
-};
-
-const userNickname = getUserNickname(); // 사용자 닉네임 가져오기
-const CookieAvatar: React.FC<{
-  responseText?: string;
-  isReceivingAudio: boolean;
-  waveform: number[];
-  isActive: boolean;
-}> = ({ responseText, isReceivingAudio, waveform, isActive }) => (
-  <View
-    style={{
-      borderColor: 'green',
-      flex: 1,
-      alignItems: 'center',
-      justifyContent: 'space-evenly',
-    }}>
-    <View style={{ flexDirection: 'column', alignItems: 'center' }}>
-      {/* 쿠키 이미지와 애니메이션을 함께 배치 */}
-      <View
-        style={{
-          width: 250,
-          height: 250,
-          alignItems: 'center',
-          justifyContent: 'center',
-          borderColor: 'purple',
-        }}>
-        {/*<AudioVisualizer isReceivingAudio={isReceivingAudio} waveform={waveform} isActive={true} />
-         앞에 위치할 쿠키 이미지 */}
-        <Text
-          style={{
-            color: 'white',
-            paddingBottom: 10,
-            fontSize: 17,
-            fontFamily: 'Pretendard-SemiBold',
-          }}>
-          리마인드 쿠키
-        </Text>
-        <View
-          style={{
-            backgroundColor: 'white',
-            width: 123,
-            height: 123,
-            borderRadius: 100,
-            overflow: 'hidden',
-            justifyContent: 'center',
-            alignItems: 'center',
-            zIndex: 1,
-          }}>
-          <Image
-            source={require('../../../src/assets/images/callcookie.png')}
-            style={{ width: 140, height: 120 }}
-          />
-        </View>
-      </View>
-    </View>
-    <View style={{ borderColor: 'pink', width: 310, height: 17 * 8 }}>
-      <ScrollView
-        style={{ height: 17 * 5 + 10 }}
-        contentContainerStyle={{
-          flexGrow: 1,
-          justifyContent: 'center',
-          alignItems: 'center',
-          paddingHorizontal: 24,
-        }}>
-        <Text
-          style={{
-            color: 'white',
-            fontFamily: 'Kyobo-handwriting',
-            fontSize: 17,
-            textAlign: 'center',
-            lineHeight: 24,
-          }}>
-          {responseText ||
-            `찾아와줘서 고마워요, ${userNickname ? `${userNickname}님` : ''}\n마음 속의 생각을 편하게 이야기 해 주세요`}
-        </Text>
-      </ScrollView>
-    </View>
-  </View>
-);
-
+import CallTimer from './components/CallTimer';
+import CookieAvatar from './components/CookieAvatar';
+import PaymentModal from './components/PaymentModal';
 const CallControls: React.FC<{
   canStart: boolean;
   canPause: boolean;
@@ -580,7 +280,9 @@ const CallPage: React.FC = () => {
       {/* 결제 모달 */}
       <PaymentModal
         visible={isPaymentModalVisible}
-        onClose={() => setIsPaymentModalVisible(false)}
+        onClose={() => {
+          setIsPaymentModalVisible(false);
+        }}
         onPayment={handlePayment}
       />
     </>
