+++ src/pages/CallPage/components/CallControls.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { View, Text } from 'react-native';
+const CallControls = () => {
+  return (
+    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+      <Text style={{ fontSize: 20, color: 'black' }}>Call Controls</Text>
+    </View>
+  );
+};
+export default CallControls;
