+++ src/pages/home/Home.tsx
@@ -303,62 +303,3 @@ const Home: React.FC<any> = ({ navigation }) => {
 };
 
 export default Home;
-
-//헤더 아이콘 클릭했을 때 이동 페이지
-/*const handleDangerPress = () => {
-    if (riskStatus === 'danger') {
-      Analytics.clickDangerLetterButton(riskScore);
-      const letterIndex = Math.floor(Math.random() * DANGER_LETTER.length);
-      setRiskData({
-        timestamp: new Date().getTime(),
-        isRead: true,
-        letterIndex,
-      });
-      navigation.navigate(RootStackName.DangerStackNavigator, {
-        screen: DangerStackName.DangerAlert,
-        params: { letterIndex },
-      }); //쿠키 편지 화면으로 이동한다
-      return;
-    }
-    if (riskStatus === 'danger-opened') {
-      //위험한 상태일 때 확인을 했으면
-      Analytics.clickOpenedDangerLetterButton(riskScore);
-      const letterIndex = getRiskData()?.letterIndex;
-      navigation.navigate(RootStackName.DangerStackNavigator, {
-        screen: DangerStackName.DangerAlert,
-        params: { letterIndex: letterIndex ?? 0 },
-      }); //쿠키 편지 화면으로 이동한다
-      return;
-    }
-    if (riskStatus === 'safe') {
-      //상담 기관 안내
-      Analytics.clickClinicInfoButton(riskScore);
-      WebBrowser.openBrowserAsync(
-        'https://autumn-flier-d18.notion.site/1268e75d989680f7b4f2d63d66f4a08a?pvs=4',
-      );
-    }
-  };*/
-
-//api 호출을 하여 위험 점수를 갱신하는 함수
-/*const refreshRiskScore = () => {
-    const date = getKoreanServerTodayDateString(new Date());
-    getRiskScore(date).then((res) => {
-      setRiskScore(res); //점수를 저장
-      if (res >= RISK_SCORE_THRESHOLD && !getRiskData()) {
-        setRiskData({
-          timestamp: new Date().getTime(),
-          isRead: false,
-          letterIndex: null,
-        });
-      }
-      refreshRiskStatus();
-    });
-  };*/
-
-//점수를 불러와서 "위험 상태"를 갱신함
-/*const refreshRiskStatus = () => {
-    const riskData = getRiskData();
-    if (!riskData) setRiskStatus('safe');
-    else if (riskData.isRead) setRiskStatus('danger-opened');
-    else setRiskStatus('danger');
-  };*/
